직관 등록 폼 컴포넌트화
